Slight modification to shopify and woocommerce platforms

+++ src/tracker-config/platforms/shopify.ts
@@ -4,14 +4,13 @@ export default function shopifyTracker(context: EcommerceContext) {
   const { appId, retailId } = context;
   const transaction = window.Shopify.checkout;
   const products = transaction.line_items;
-  // window.Shopify.checkout.order.total_price * 100;
 
   window.tracker("addTrans",
     transaction.order_id.toString(),
     !retailId ? appId : retailId,
-    parseInt(transaction.total_price),
-    parseInt(transaction.total_tax ? transaction.total_tax : 0),
-    parseInt(transaction.shipping_rate.price ? transaction.shipping_rate.price : 0),
+    parseFloat(transaction.total_price),
+    parseFloat(transaction.total_tax ? transaction.total_tax : 0),
+    parseFloat(transaction.shipping_rate.price ? transaction.shipping_rate.price : 0),
     (transaction.billing_address.city ?  transaction.billing_address.city : "N/A").toString(),
     (transaction.billing_address.province ? transaction.billing_address.province : "N/A").toString(),
     (transaction.billing_address.country ?  transaction.billing_address.country : "N/A").toString(),
@@ -24,8 +23,8 @@ export default function shopifyTracker(context: EcommerceContext) {
       products[i].id.toString(),
       (products[i].title ? products[i].title : "N/A").toString(),
       (products[i].variant_title ? products[i].variant_title : "N/A").toString(),
-      parseInt(products[i].price),
-      parseInt(products[i].quantity ? products[i].quantity : 1),
+      parseFloat(products[i].price),
+      parseFloat(products[i].quantity ? products[i].quantity : 1),
       transaction.currency.toString()
     );
   }
